refactorings to make l10n infrastructure easier to understand.
